sharpclaw: drop the (s32) cast — it mints an instruction retail lacks
Reverts the cast I landed earlier today, for the same reason the player.c
sites were corrected: the cast emits srawi r0,r3,31 immediately before
retail's li r0,-65; and. Without it our count is exactly retail's
250/250 and jtoracle reports zero divergences for the unit.

Function fuzzy 99.600 -> 99.156 with matched_code, matched_data and
matched_functions all unchanged — the (s32) form scored HIGHER fuzzy
while emitting an instruction retail does not have. That is precisely
the trap jtoracle exists to catch.

+++ src/dlls/objects/202/sharpclaw.c
@@ -586,7 +586,7 @@ u32 wispBaddieProcessAnimEvent(GameObject* obj, u8* state, u32 allowNewEvent)
         if (((EnemyState*)state)->sharpClaw.moveHoldTimer <= 0.0f)
         {
             ((EnemyState*)state)->sharpClaw.moveHoldTimer = 0.0f;
-            ((EnemyState*)state)->controlFlags = (s32)((EnemyState*)state)->controlFlags & ~0x40LL;
+            ((EnemyState*)state)->controlFlags = ((EnemyState*)state)->controlFlags & ~0x40LL;
             ((EnemyState*)state)->controlFlags =
                 ((EnemyState*)state)->controlFlags | (u64)BADDIE_CONTROL_SEQUENCE_DRIVEN;
             ((EnemyState*)state)->curveIndex = ((EnemyState*)state)->curveIndex & ~0x80;
